refactor(frontend): create useNavigationHelpers hook, deduplicate handleGoBack/handleGoHome
